Rename thapbi_pict plate-summary to read-summary
This reflects the report being unique sequence or read centric,
with unique sequences as the rows and their read counts as the
main data values.

Also, if we add a two-dimensional 8x12 report for use with 96 well
plates, that would make more sense to name as plate-summary.
